Remove old EIA-860 and EIA-923 output modules.
